add a util to process registry url
