+++ src/sync.rs
@@ -1,10 +1,10 @@
+use crate::util::Canonicalized;
 use crate::{util, Krate, Source};
 use anyhow::{Context, Error};
 use std::{convert::TryFrom, io::Write, path::PathBuf};
 use tracing::{debug, error, info, warn};
 use tracing_futures::Instrument;
 
-pub const CRATES_IO_URL: &str = "https://github.com/rust-lang/crates.io-index";
 pub const INDEX_PATH: &str = "registry/index";
 pub const INDEX_DIR: &str = "registry/index/github.com-1ecc6299db9ec823";
 pub const CACHE_DIR: &str = "registry/cache/github.com-1ecc6299db9ec823";
@@ -15,16 +15,9 @@ pub const GIT_CO_DIR: &str = "git/checkouts";
 pub async fn registry_index(
     root_dir: PathBuf,
     backend: crate::Storage,
-    registry_url: Option<util::Canonicalized>,
+    registry_url: Option<Canonicalized>,
 ) -> Result<(), Error> {
-    let canonicalized = match registry_url {
-        Some(u) => u,
-        None => {
-            let u = url::Url::parse(CRATES_IO_URL)?;
-            util::Canonicalized::try_from(&u)?
-        }
-    };
-    let ident = canonicalized.ident();
+    let (canonicalized, ident) = util::decode_registry_url(registry_url)?;
 
     let index_path = root_dir.join(INDEX_PATH).join(ident.clone());
     std::fs::create_dir_all(&index_path).context("failed to create index dir")?;
